refractor: add seperate printer
